Move unit test for meansquared into test_energy module
